fix(ingest-plan): cast numerics so the plan can be JSON-serialized
plan_ingest raised "Object of type Decimal is not JSON serializable" on every call,
including a three-word payload. Reported by Chat with two request IDs and a correct
minimal repro.

Cause: Postgres `round(extract(epoch ...))` returns numeric, which psycopg maps to
Decimal, and Decimal has no JSON encoder. `age_days` was the field that fired;
`similarity` from the pgvector expression had the same exposure and is fixed with it.
The plan object built correctly in Python — which is why the live check during
development passed — and only failed when an MCP surface serialized it to the wire.

Fixed at the source with ::int / ::float casts rather than at the encoder, so no future
caller has to remember. Comment left at both sites explaining the casts are load-bearing.

Root cause of the miss: all 14 existing tests stub build_plan, so the DB -> JSON path had
zero coverage. A stub cannot produce a type that only exists on the wire. Added two
DB-backed tests (skipped without SUPABASE_DB_URL): one asserts the real plan survives
json.dumps, one asserts age_days is int and similarity is float — because json.dumps
alone would still pass if a field were silently stringified instead of cast.

Verified against Chat's exact minimal repro (source="Test.", system=FlightSim,
component=flightsim-hardware): serializes clean, would_supersede correctly resolves to
b5041d45. 164 passed, 1 xfailed. ruff clean.

+++ api/ingest_plan.py
@@ -115,8 +115,12 @@ def _living_docs(conn, system: str | None) -> list[dict[str, Any]]:
     if not system:
         return []
     rows = conn.execute(
+        # ::int / ::float casts are load-bearing, not cosmetic. Postgres `round()` over
+        # `extract(epoch ...)` returns numeric, which psycopg maps to Decimal, and Decimal is not
+        # JSON-serializable — so the plan built fine in Python and then 500'd the moment an MCP
+        # surface tried to serialize it. Cast at the source so no caller has to remember.
         """SELECT id::text, system, component_key,
-                  round(extract(epoch FROM (now() - created_at)) / 86400) AS age_days,
+                  round(extract(epoch FROM (now() - created_at)) / 86400)::int AS age_days,
                   split_part(regexp_replace(content, '^#+\\s*', ''), E'\\n', 1) AS title
              FROM public.knowledge
             WHERE system = %s AND component_key IS NOT NULL AND status = 'current'
@@ -134,8 +138,9 @@ def _similar_docs(conn, embedding_sql_param, limit: int = 3) -> list[dict[str, A
     if embedding_sql_param is None:
         return []
     rows = conn.execute(
+        # ::float for the same reason as _living_docs — see the note there.
         """SELECT id::text, system, component_key,
-                  1 - (embedding <=> %s::vector) AS similarity,
+                  (1 - (embedding <=> %s::vector))::float AS similarity,
                   split_part(regexp_replace(content, '^#+\\s*', ''), E'\\n', 1) AS title
              FROM public.knowledge
             WHERE component_key IS NOT NULL AND status = 'current' AND embedding IS NOT NULL
